bft_network_partitioning: modify to support TCP communication
Until this change, bft_network_partitioning  implementation is very
simple - it knows in advance source and destination ports.
This is not the case when dealing with TCP communication, where the TCP
client (which might be a BFT network client or replica) gets an
unknown port from the TCP server.
To achieve an easy port mapping we relay on 2 facts:
1) All network members have already connected. They must be started
before opening an advisory context.
2) If connections are changed, they are not changed during isolation
time - that means, ports are the same for the life of the isolation.
Nodes may be started or stopped only before or after the lifetime of
the advisory.

It is easier to use shell commands and parse their output than using
native python packages. We use fuser, netstat, and grep to parse
match the node ids into ports. We also use data stored in clients and in
BFT network to reach our goal.