BCStateTran: Add a periodic status log while in GettingMissingBlocks
There was very little information, till now, on the progress of the
state transfer's missing blocks collection process (state
GettingMissingBlocks).

Here we add a new class Throughput, which helps tracking the number of
bytes and blocks processed in a time unit. This class may be used later
(with some modifications) for any process which requires
throughput measurements.

The time duration which is used to measure the throughput might be an
overall time since collection began, or last reporting window (which is
for now a half window - 150 blocks, which is equal for a checkpoint
time).

When a replica starts collecting, it calls startCollectingStats(), which
calls start() on all statistics being collected.
Every time a block is collected - report() is called to report a block
has been collected, and it's size. Every 150 blocks a summary message is
logged by calling logGettingMissingBlocksSummary(). Pretty formatting
is not used, to support Log Insight processing (one-liner).
The function also prints recent checkpoints information,
as the network might progress even during the state transfer.