BCStateTran: Add latency histograms to dst send messages
Here we add 3 types of histograms in order to measure latency for
3 type of messages send from dst:
1) AskForCheckpointSummariesMsg
2) FetchBlocksMsg
3) FetchResPagesMsg

msg_latency_tracker_ holds pairs of message sequence number and time
message send. Upon receiving reply to that sequence number, delta time
is calculated and inserted to the matching histogram.
snapshot is taken and all histograms are printed to logs when ST ends.