BftTestNetwork changes for client transaction signing
If client transaction signing is enabled in test environment, these changes will setup
the two parameters that the tester replica needs:
-p for the principals mapping string.
-t for the transaction signing keys path.
and then passes these parameters to the tester replica.

This will also pass the transaction signing keys path and principals to participants map
to the bft client through the bft config.