Reduce copying when primary sends to all replicas
Prior to this change, any time the primary sent to all replicas it would
make a copy of the message and then call `send` for each individual
replica. This resulted in a copy, as well as an additional call to the
communication layer.

The communication layer already optimizes out the copying at the upper
layer. This commit plumbs through the interface changes required to use
that optimization. It then proceeds to use the optimization via
`sendToAllOtherReplicas` in ReplicaImp.

This is particularly important when sending PrePrepare messages, as
those can get quite large due to consensus batching. This change
improved latency considerably in our performance tests with 25kb
requests and batches of 100 (2.5MB total size). We now only copy 2.5MB
once instead of 6 times.

```
Count: 40
Max: 20687
Min: 459
Percentiles:
  10: 3543
  25: 4763
  50: 7479
  75: 11127
  90: 17135
  95: 19567
  99: 20687
  99.9: 20687
  99.99: 20687
  99.999: 20687
```

```
Count: 63
Max: 4019
Min: 293
Percentiles:
  10: 300
  25: 335
  50: 529
  75: 982
  90: 2575
  95: 3077
  99: 4001
  99.9: 4019
  99.99: 4019
  99.999: 4019
```