BCStateTran: intro BlockIODataPool, refactor getBlock at source, fix bug
We would like to use the same memory allocated to srcGetBlockContextes_
in both source (get block) and destination (put block), and later on
we might need it for other uses.
3 main changes:
1) Introduce a very basic memory pool, holding pre-allocated
shared pointers to BlockIOContext. it support few basic operation, the
important ones are allloc and free. It throws an exception if alloc
fails. On free, it waits for the future is job is not yet finished.
2) Convert buffer_ raw pointer into smart pointer.
3) As a result of the change in 1, we can now simplify, refactor and
adjust the code in source :
BCStateTran::onMessage(const FetchBlocksMsg ..
during refactoring I've also found a rare bug while sending blocks
in chunks.
4) To keep uniform naming convention:
rename asyncGetBlocksConcurrent to getBlocksConcurrentAsync